changed command base to help base which now handels only the help features of a command
